Split widget: glass card
